Add 'qc' command
Keep the name 'qc' or rename to 'proper' as suggested by Kostis?
Kostis suggested we might want to have a way to ask rebar to test PropEr
or EQC or FooQC properties specifically with that qc lib. I did add an
option to specify a qc lib to be used for all properties, but Kostis was
suggesting we might need to have such an option or separate commands
(rebar proper, rebar eqc, ....) for different parts of the source tree
to test properties in.